refactor : 패키지 명, 도메인에 맞게 변경
