--- src/main/java/nextstep/ladder/domain/participant/Participant.java
@@ -1,34 +1,0 @@
-package nextstep.ladder.domain.participant;
-
-import java.util.Objects;
-import java.util.Optional;
-
-public class Participant {
-    private final Name name;
-
-    public Participant(Name name) {
-        this.name = Optional.ofNullable(name)
-                .orElseThrow(() -> new IllegalArgumentException("이름은 빈 칸일 수 없습니다."));
-    }
-
-    public boolean isParticipant(String name) {
-        return this.name.isName(name);
-    }
-
-    String getName() {
-        return this.name.getName();
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        Participant that = (Participant) o;
-        return Objects.equals(name, that.name);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(name);
-    }
-}
