--- src/main/java/nextstep/ladder/vo/MoveStrategy.java
@@ -1,6 +1,0 @@
-package nextstep.ladder.vo;
-
-@FunctionalInterface
-public interface MoveStrategy {
-    int move();
-}
