+++ src/main/java/nextstep/laddergame/domain/reward/Reward.java
@@ -1,0 +1,28 @@
+package nextstep.laddergame.domain.reward;
+
+import java.util.Objects;
+
+public class Reward {
+    private final String reward;
+
+    public Reward(String reward) {
+        this.reward = reward;
+    }
+
+    public String getReward() {
+        return reward;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        Reward reward1 = (Reward) o;
+        return Objects.equals(reward, reward1.reward);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(reward);
+    }
+}
