+++ src/main/java/nextstep/laddergame/vo/MoveStrategy.java
@@ -1,0 +1,6 @@
+package nextstep.laddergame.vo;
+
+@FunctionalInterface
+public interface MoveStrategy {
+    int move();
+}
